+++ system/main.cpp
@@ -159,8 +159,8 @@ int main(int argc, char* argv[])
 
 		/* Lane estimation */
 		float d = 0, phi = 0;
-		vector<Vec4i> outer_lines;
-		vector<Vec4i> inner_lines;
+		vector<Vec4f> outer_lines;
+		vector<Vec4f> inner_lines;
 		lane_detector->lane_detect(distort_image, outer_lines, inner_lines);
 		bool get_pose = lane_detector->pose_estimate(inner_lines, d, phi);
 		lane_detector->publish_images();
